refactor(Styled): main layout and leaderboard components
