+++ src/layout/main/app-bar/app-bar.tsx
@@ -1,12 +1,8 @@
 import { useState } from 'react';
 import type { MouseEvent } from 'react';
-
 import { Container, Toolbar, AppBar as AppBarMui } from '@mui/material';
-import { DesktopLogo } from 'layout/main/app-bar/desktop-logo';
-import { MobileLogo } from 'layout/main/app-bar/mobile-logo';
-import { User } from 'layout/main/app-bar/user';
-import { MobileMenu } from './mobile-menu';
-import { DesktopMenu } from './desktop-menu';
+
+import { DesktopLogo, User, MobileLogo, MobileMenu, DesktopMenu } from './components';
 
 const pages = ['Лидеры', 'Форум'];
