+++ src/layout/main/app-bar/components/index.ts
@@ -1,0 +1,5 @@
+export { DesktopLogo } from './desktop-logo';
+export { DesktopMenu } from './desktop-menu';
+export { MobileLogo } from './mobile-logo';
+export { MobileMenu } from './mobile-menu';
+export { User } from './user';
