--- src/layout/main/app-bar/mobile-logo.tsx
@@ -1,26 +1,0 @@
-import { FlutterDash as AppIcon } from '@mui/icons-material';
-import { Typography } from '@mui/material';
-
-export const MobileLogo = () => (
-  <>
-    <AppIcon sx={{ display: { xs: 'flex', md: 'none' }, mr: 1 }} />
-    <Typography
-      variant="h5"
-      noWrap
-      component="a"
-      href=""
-      sx={{
-        mr: 2,
-        display: { xs: 'flex', md: 'none' },
-        flexGrow: 1,
-        fontFamily: 'monospace',
-        fontWeight: 700,
-        letterSpacing: '.3rem',
-        color: 'inherit',
-        textDecoration: 'none',
-      }}
-    >
-      Flappy Bird
-    </Typography>
-  </>
-);
