+++ src/layout/main/main.tsx
@@ -6,7 +6,7 @@ import { AppBar } from './app-bar';
 export const MainLayout: FC = ({ children }) => (
   <>
     <AppBar />
-    <Container fixed sx={{ paddingTop: 1 }}>
+    <Container fixed sx={{ pt: 1 }}>
       {children}
     </Container>
   </>
